feat(UIKIT-1475,NewDataGrid): Правки со storybook и типизации
