+++ packages/components/src/NewDataGrid/faker.ts
@@ -1,6 +1,12 @@
 import { faker } from '@faker-js/faker';
 
-import { type DataGridRowOptionColumns } from '../NewDataGrid';
+import type {
+  CellValue,
+  DataGridColumns,
+  DataGridRowOptionColumns,
+  DataGridRowOptions,
+  DataGridRowWithOptions,
+} from './types';
 
 type DataType = {
   id: string;
@@ -22,13 +28,46 @@ const makeRandomDate = () => {
   return randomDate.toISOString();
 };
 
-export const makeDataList = (length: number = 10): DataType[] => {
-  return Array.from({ length }).map((_, i) => ({
-    id: faker.string.uuid(),
-    documentName: `Договор №${i + 1}`,
-    recipient: RECIPIENTS[Math.floor(Math.random() * RECIPIENTS.length)],
-    createDate: makeRandomDate(),
+export const makeColumns = (
+  columnsTemplate: DataGridColumns<DataType>[],
+  mergedColumns: DataGridColumns<DataType>[] = [],
+): DataGridColumns<DataType>[] => {
+  const mergedColumnsMap = mergedColumns.reduce(
+    (acc, { field, ...columnsOptions }) => {
+      if (field) {
+        return {
+          ...acc,
+          [field]: columnsOptions,
+        };
+      }
+
+      return acc;
+    },
+    {},
+  );
+
+  return columnsTemplate.map((column) => ({
+    ...column,
+    // @ts-ignore
+    ...(mergedColumnsMap[column?.field] || {}),
   }));
+};
+
+export const makeDataList = <TData extends Record<string, CellValue>>(
+  dataObjTemplate: TData,
+  options?: DataGridRowOptions<TData>,
+): DataGridRowWithOptions<TData>[] => {
+  const DATA_ARRAY_LENGTH = 16;
+
+  return Array.from({ length: DATA_ARRAY_LENGTH })
+    .fill(dataObjTemplate)
+    .map((_, i) => ({
+      id: String(i + 1),
+      documentName: `Договор №${i + 1}`,
+      recipient: RECIPIENTS[Math.floor(Math.random() * RECIPIENTS.length)],
+      createDate: makeRandomDate(),
+      options: Math.random() < 0.5 ? options : {},
+    })) as unknown as DataGridRowWithOptions<TData>[];
 };
 
 export const makeDataListWithTree = (
